优化下线后获取服务列表

+++ moduleOperation/service.js
@@ -464,7 +464,7 @@ service.uninstallModule4DCache = async function ({ unType = 2, appName, moduleNa
   const { __return, uninstallRsp, uninstallRsp: { errMsg } } = await DCacheOptPrx.uninstall4DCache(option);
   assert(__return === 0, errMsg);
   const module = await ModuleConfigService.getModuleConfigByName({ moduleName });
-  await ModuleConfigService.removeModuleConfig({ module_name: moduleName, module_id: module.module_id });
+  if (module) await ModuleConfigService.removeModuleConfig({ module_name: moduleName, module_id: module.module_id });
   await service.getUninstallPercent(option);
   return uninstallRsp;
 };
@@ -490,10 +490,10 @@ service.getUninstallPercent = async function (option) {
   const { __return, UninstallProgressRsp: { percent, errMsg } } = await DCacheOptPrx.getUninstallPercent(option);
   assert(__return === 0, errMsg);
   if (+percent !== 100) {
-    uninstallTimer = setTimeout(() => {
-      service.getUninstallPercent(option);
-    }, 1000);
-  } else if (uninstallTimer) clearTimeout(uninstallTimer);
+    uninstallTimer = setTimeout(service.getUninstallPercent.bind(this, option), 500);
+  } else {
+    clearTimeout(uninstallTimer);
+  }
 };
 
 module.exports = service;
